Refactor out differential drive functionality to a separate class.
